chatbot-service/mcp_server/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
loguru>=0.7.2
httpx>=0.25.1
python-dotenv>=1.0.0